+++ webapp/components/rhs_root_post.jsx
@@ -359,7 +359,7 @@ export default class RhsRootPost extends React.Component {
                                     <FormattedDate
                                         value={post.create_at}
                                         day='numeric'
-                                        month='long'
+                                        month='short'
                                         year='numeric'
                                         hour12={!this.props.useMilitaryTime}
                                         hour='2-digit'
